Fixed implementation of datastructures
